Moved CursorDataExtractor and ContentValuesPopulator to package datastore since the sub packages now only had the interfaces.
Also removed from DataStore interface accessors for both since this is only an implementation detail.
