+++ src/java/com/sarality/app/datastore/AbstractContentResolverDataStore.java
@@ -8,7 +8,6 @@ import android.database.Cursor;
 import android.net.Uri;
 import android.util.Log;
 
-import com.sarality.app.datastore.extractor.CursorDataExtractor;
 import com.sarality.app.datastore.query.Query;
 
 public abstract class AbstractContentResolverDataStore<T> extends AbstractDataStore<T> implements DataStore<T> {
