+++ tests/src/com/sarality/app/datastore/db/test/SqliteTableTest.java
@@ -23,12 +23,12 @@ import com.sarality.app.common.collect.Lists;
 import com.sarality.app.datastore.Column;
 import com.sarality.app.datastore.ColumnDataType;
 import com.sarality.app.datastore.ColumnSpec;
+import com.sarality.app.datastore.ContentValuesPopulator;
+import com.sarality.app.datastore.CursorDataExtractor;
 import com.sarality.app.datastore.db.SqliteTable;
 import com.sarality.app.datastore.db.SqliteTableSchemaUpdater;
 import com.sarality.app.datastore.db.TableColumnProperty;
 import com.sarality.app.datastore.db.TableInfo;
-import com.sarality.app.datastore.extractor.CursorDataExtractor;
-import com.sarality.app.datastore.populator.ContentValuesPopulator;
 import com.sarality.app.datastore.query.Query;
 
 /**
