+++ tests/src/com/sarality/app/datastore/db/test/TestTable.java
@@ -5,10 +5,10 @@ import java.util.List;
 import android.app.Application;
 
 import com.sarality.app.datastore.Column;
+import com.sarality.app.datastore.ContentValuesPopulator;
+import com.sarality.app.datastore.CursorDataExtractor;
 import com.sarality.app.datastore.db.SqliteTable;
 import com.sarality.app.datastore.db.SqliteTableSchemaUpdater;
-import com.sarality.app.datastore.extractor.CursorDataExtractor;
-import com.sarality.app.datastore.populator.ContentValuesPopulator;
 
 class TestTable extends SqliteTable<TestObject> {
   protected TestTable(Application application, String dbName, String tableName, int tableVersion,
