+++ tests/src/com/sarality/app/datastore/test/AbstractContentResolverDataStoreTest.java
@@ -17,8 +17,8 @@ import android.test.mock.MockContentResolver;
 import com.sarality.app.common.collect.Lists;
 import com.sarality.app.datastore.AbstractContentResolverDataStore;
 import com.sarality.app.datastore.Column;
+import com.sarality.app.datastore.CursorDataExtractor;
 import com.sarality.app.datastore.db.test.TestObject;
-import com.sarality.app.datastore.extractor.CursorDataExtractor;
 import com.sarality.app.datastore.query.Query;
 
 /**
