+++ tests/src/com/sarality/app/datastore/test/AbstractDataStoreTest.java
@@ -13,8 +13,8 @@ import org.robolectric.RobolectricTestRunner;
 import com.sarality.app.common.collect.Lists;
 import com.sarality.app.datastore.AbstractDataStore;
 import com.sarality.app.datastore.Column;
+import com.sarality.app.datastore.CursorDataExtractor;
 import com.sarality.app.datastore.db.test.TestObject;
-import com.sarality.app.datastore.extractor.CursorDataExtractor;
 import com.sarality.app.datastore.query.Query;
 
 /**
