+++ tests/src/com/sarality/app/datastore/test/AbstractWritableDataStoreTest.java
@@ -13,9 +13,9 @@ import org.robolectric.RobolectricTestRunner;
 import com.sarality.app.common.collect.Lists;
 import com.sarality.app.datastore.AbstractWritableDataStore;
 import com.sarality.app.datastore.Column;
+import com.sarality.app.datastore.ContentValuesPopulator;
+import com.sarality.app.datastore.CursorDataExtractor;
 import com.sarality.app.datastore.db.test.TestObject;
-import com.sarality.app.datastore.extractor.CursorDataExtractor;
-import com.sarality.app.datastore.populator.ContentValuesPopulator;
 import com.sarality.app.datastore.query.Query;
 
 /**
